Fix processing annotations set in kotlin classes
Fix the unresolved `TypeAdapter` annotation that is set in a kotlin
class.

This seems to happen because `kapt` generates java stubs
for annotation processor and these stubs miss the annotations with
`RetentionPolicy.SOURCE`. So if you need to access an element that was
not loaded by the `RoundEnvironment` at some point, you'll have `null`
instead of an annotation instance even if it is actually present there.

Apparently the `kapt` issue (or a limitation) is not described well, but
mentioned in the [kotlin blog post](https://blog.jetbrains.com/kotlin/2015/06/better-annotation-processing-supporting-stubs-in-kapt/)
and the similar issue experienced the developers of [PermissionDispatcher](https://github.com/permissions-dispatcher/PermissionsDispatcher/issues/112).

So the workaround here is to use `RetentionPolicy.CLASS` instead so the
annotation remain in the generated java stub during all processing
rounds, but it still won't be loaded by the class loader on runtime.