Add @DefaultValue annotation for path segments and query params
- Add a new annotation `@DefaultValue` that can be used as a source of a
fallback value for nullable method parameters.
- Update `Intependent Uri`, `Associated Uri` and `Uri Factory` generator
steps to be able to read the default value from the new annotation.
- Adjust `TypeConverter` to be able to use the default value as a
fallback for nullable method parameters, absent query parameters or
invalid data in the query parameters or path segments (eg: a random
string received in the `Uri` query parameter for a field of an integer
type).
- Adjust `AnnotationHelper` to be able to replace `@Nullable` annotation
of a getter method with `@NonNull` in case `@DefaultValue` is specified.
- Minor fixes and adjustments.

Resolves #9