API: Add ability to generate UriMatcher for UriFactory
- Add `WithUriMatcher` annotation to specify which `UriFactory` can have
a `UriMatcher`.
- Add `MatchesTo` annotation to specify the matcher code constant name
and generate unique matcher code values so they never repeat across the
whole app.
- Add `MatcherCode` annotation to specify the matcher code value in case
the uniqueness of the matcher codes is handled by the app developers.
- Add `Element.getAnnotation()` and `Element.requireAnnotation()`
kotlin extensions in order to simplify the code readability.
- Add `UriMatcherGeneratorStep` to `UriFactoryProcessor` in order to be
able to generate a `UriMatcher` implementation for a specific
`UriFactory`.
- Minor fixes in `IndependentUriDataProcessor`,
`TypeAdapterFactoryGeneratorStep` and `UriFactoryGeneratorStep`.

Resolves #15