Change common modules' dependency layout
In order to be able to publish a fat jar with all common classes packed
in a single jar the top-level processor modules (`processor` and `processor-ksp`) should
include all the dependencies as `implementation` while the common
modules have them as `compileOnly`. This way a `pom` file for the
artifact will contain only a list of 3d party dependencies while the
classes from the common modules will be included in the jar of the
top-level module.